refactor org pending pages
