--- libs/e2e/src/lib/pages/organization/OrganizationJoinPendingPage.ts
@@ -1,15 +1,0 @@
-import { SEC } from '../../utils/env';
-
-const PATH = '/c/organization/join-congratulations';
-
-export default class OrganizationJoinPendingPage {
-  constructor() {
-    cy.get('organization-pending', {timeout: 150 * SEC});
-  }
-
-  public assertMoveToOrgJoinPage() {
-    cy.location().should(loc => {
-      expect(loc.pathname).to.eq(PATH);
-    });
-  }
-}
